End turn when no enemies left, menu music fix
Menu music was continuing from where it left off after a level. Note
that because of the way it's checking for scene name we'll have to
update MainTheme if we want to have a third level

+++ Assets/Level/LevelScript.cs
@@ -226,9 +226,18 @@ public class LevelScript : MonoBehaviour
             script.greyOutMoveButton();
         }
         //Cheat code: Remove enemy to progress to next level 
-        /*if (Input.GetKeyDown("space"))
+        if (Input.GetKeyDown("space"))
             enemies.RemoveRange(0, 1);
-        */
+        
+
+        for (int i = 0; i < enemies.Count; i++)
+        {
+            if (enemies[i].getState() == EnemyBehaviour.EnemyState.Dead)
+            {
+                enemies.Remove(enemies[i]);
+            }
+        }
+
         if (enemies.Count <= 0)
         {
             manager.level++;
